Remove multi-arch structure
