+++ lib/commands/install.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const install = require('../install/macos');
+const installation = require('../installation');
 
 module.exports = /** @type {import('yargs').CommandModule} */ ({
   command: 'install',
@@ -12,6 +12,6 @@ module.exports = /** @type {import('yargs').CommandModule} */ ({
     });
   },
   async handler({ unattended }) {
-    await install.install({ unattended });
+    await installation.install({ unattended });
   },
 });
