+++ lib/commands/uninstall.js
@@ -1,12 +1,12 @@
 'use strict';
 
-const install = require('../install/macos');
+const installation = require('../installation');
 
 module.exports = /** @type {import('yargs').CommandModule} */ ({
   command: 'uninstall',
   describe: 'Uninstall text to speech extension and other support',
   async handler() {
-    await install.uninstall();
+    await installation.uninstall();
 
     console.log('Uninstallation completed successfully.');
   },
