+++ lib/helpers/Modifier.js
