+++ lib/helpers/applescript.js
