+++ lib/helpers/keyCodeCommand.js
