+++ lib/helpers/parseCodePoints.js
@@ -1,4 +1,4 @@
-/// <reference path="../../modules/types.js" />
+/// <reference path="../modules/types.js" />
 /// <reference path="./applescript.js" />
 
 'use strict';
