+++ lib/helpers/types.js
