+++ lib/helpers/validateCommand.js
