+++ lib/installation.js
@@ -7,7 +7,7 @@ const { promisify } = require('util');
 
 const debug = require('debug')('install');
 
-const { 'interaction.userIntent': userIntent } = require('../modules/macos/interaction');
+const { 'interaction.userIntent': userIntent } = require('./modules/interaction');
 
 const LSREGISTER_EXECUTABLE_PATH =
   '/System/Library/Frameworks/CoreServices.framework/Versions/Current/Frameworks/LaunchServices.framework/Versions/Current/Support/lsregister';
