+++ lib/modules/interaction.js
@@ -2,6 +2,19 @@
 
 'use strict';
 
-const interaction = require('./macos/interaction');
+const { runScript, renderScript } = require('../helpers/applescript');
+const { parseCodePoints } = require('../helpers/parseCodePoints');
 
-module.exports = interaction;
+const userIntent = /** @type {ATDriverModules.InteractionUserIntent} */ (
+  async function (websocket, { name, keys }) {
+    if (name !== 'pressKeys') {
+      throw new Error('unknown user intent');
+    }
+    await runScript(renderScript(parseCodePoints(keys)));
+    return {};
+  }
+);
+
+module.exports = /** @type {ATDriverModules.Interaction} */ ({
+  'interaction.userIntent': userIntent,
+});
