+++ src/MacOSATDriverServer/MacOSATDriverServerExtension/Info.plist
@@ -15,7 +15,7 @@
 						The "manufacturer" must be kept in-sync with other
 						references in this project:
 
-						- lib/install/macos.js
+						- lib/installation.js
 						- src/MacOSATDriverServer/MacOSATDriverServer/Model/AudioUnitHostModel.swift
 					-->
 					<key>manufacturer</key>
@@ -28,7 +28,7 @@
 						The "subtype" must be kept in-sync with other
 						references in this project:
 
-						- lib/install/macos.js
+						- lib/installation.js
 						- src/MacOSATDriverServer/MacOSATDriverServer/Model/AudioUnitHostModel.swift
 					-->
 					<key>subtype</key>
@@ -41,7 +41,7 @@
 						The "type" must be kept in-sync with other references
 						in this project:
 
-						- lib/install/macos.js
+						- lib/installation.js
 						- src/MacOSATDriverServer/MacOSATDriverServer/Model/AudioUnitHostModel.swift
 					-->
 					<key>type</key>
